Enabling User File Access Sharing (#1251)
### Share File
Implemented the feature of UserFile Access Sharing
- `UserFile` now has a `AccessEntry` associated with it. The corresponding MySQL table is `user_file_access`.
- `UserFile` has two `accessLevel`s, `Read` and `Write`. It also has a special `Owner` accessLevel, which can be treated as `Write`.
    - `Read` access can execute the file.
    - `Write` access can additionally delete the file.
    - Owner of the file can decide to share the workflow to another user, with either `Read` or `Write` access.

### Shared File Execution
- Shared File are being uniquely identified as `{OwnerName}/{FileName}`, across the dashboard and workspace. <img width="333" alt="Screen Shot 2021-08-05 at 10 51 18" src="https://user-images.githubusercontent.com/17627829/128397558-c517475b-1564-4c2f-b39c-c93d340757c0.png">


- Shared file can be executed in the workflow, the actual file path is resolved from database. <img width="769" alt="Screen Shot 2021-08-05 at 10 51 32" src="https://user-images.githubusercontent.com/17627829/128397570-a2ece10f-bc0b-47e7-81e7-6dc16843f0b0.png">

### UI Update
This PR also includes some frontend UI changes:
- Now shows access entries as badges.
<img width="180" alt="Screen Shot 2021-08-05 at 10 48 54" src="https://user-images.githubusercontent.com/17627829/128397215-8aae558f-e0b1-487f-9ad7-dbef559b7a2b.png">

- Now shows  
    - an icon 
<img width="27" alt="Screen Shot 2021-08-05 at 10 47 56" src="https://user-images.githubusercontent.com/17627829/128397085-bf5fd416-e84d-4cca-b544-3b48ca30045a.png"> to indicate ownership
    - and another icon
<img width="29" alt="Screen Shot 2021-08-05 at 10 47 37" src="https://user-images.githubusercontent.com/17627829/128397120-55e77cc7-bcd9-46f8-bfea-9fc07d21281b.png"> to indicate being shared.

+++ core/amber/src/main/scala/edu/uci/ics/texera/web/resource/dashboard/file/UserFileResource.scala
@@ -1,13 +1,15 @@
 package edu.uci.ics.texera.web.resource.dashboard.file
 
 import edu.uci.ics.texera.web.SqlServer
-import edu.uci.ics.texera.web.model.jooq.generated.Tables.FILE
-import edu.uci.ics.texera.web.model.jooq.generated.tables.daos.FileDao
-import edu.uci.ics.texera.web.model.jooq.generated.tables.pojos.File
+import edu.uci.ics.texera.web.model.jooq.generated.Tables.{FILE, USER_FILE_ACCESS}
+import edu.uci.ics.texera.web.model.jooq.generated.tables.daos.{FileDao, UserDao, UserFileAccessDao}
+import edu.uci.ics.texera.web.model.jooq.generated.tables.pojos.{File, User, UserFileAccess}
 import edu.uci.ics.texera.web.resource.auth.UserResource
+import edu.uci.ics.texera.web.resource.dashboard.file.UserFileResource.context
 import io.dropwizard.jersey.sessions.Session
 import org.apache.commons.lang3.tuple.Pair
 import org.glassfish.jersey.media.multipart.{FormDataContentDisposition, FormDataParam}
+import org.jooq.DSLContext
 import org.jooq.types.UInteger
 
 import java.io.InputStream
@@ -16,17 +18,32 @@ import java.util
 import javax.servlet.http.HttpSession
 import javax.ws.rs._
 import javax.ws.rs.core.{MediaType, Response}
+import scala.collection.JavaConverters._
+import scala.collection.mutable
 
 /**
   * Model `File` corresponds to `core/new-gui/src/app/common/type/user-file.ts` (frontend).
   */
+case class DashboardFileEntry(
+    ownerName: String,
+    accessLevel: String,
+    isOwner: Boolean,
+    file: File
+)
+
+object UserFileResource {
+  private val context: DSLContext = SqlServer.createDSLContext
+}
 
 @Path("/user/file")
 @Consumes(Array(MediaType.APPLICATION_JSON))
 @Produces(Array(MediaType.APPLICATION_JSON))
 class UserFileResource {
-
-  final private val fileDao = new FileDao(SqlServer.createDSLContext.configuration)
+  final private val fileDao = new FileDao(context.configuration)
+  final private val userFileAccessDao = new UserFileAccessDao(
+    context.configuration
+  )
+  final private val userDao = new UserDao(context.configuration)
 
   /**
     * This method will handle the request to upload a single file.
@@ -65,36 +82,98 @@ class UserFileResource {
             description
           )
         )
+        val fid = context
+          .select(FILE.FID)
+          .from(FILE)
+          .where(FILE.UID.eq(userID).and(FILE.NAME.eq(fileName)))
+          .fetch()
+          .getValue(0, 0)
+          .asInstanceOf[UInteger]
+        userFileAccessDao.insert(
+          new UserFileAccess(userID, fid, true, true)
+        )
         Response.ok().build()
       case None =>
         Response.status(Response.Status.UNAUTHORIZED).build()
     }
   }
 
+  /**
+    * This method returns a list fo all files accessible by the current user
+    * @param session the session indicating current logged-in user
+    * @return
+    */
   @GET
   @Path("/list")
-  def listUserFiles(@Session session: HttpSession): util.List[File] = {
+  def listUserFiles(@Session session: HttpSession): util.List[DashboardFileEntry] = {
     UserResource.getUser(session) match {
-      case Some(user) => getUserFileRecord(user.getUid)
-      case None       => new util.ArrayList[File]()
+      case Some(user) => getUserFileRecord(user)
+      case None       => new util.ArrayList[DashboardFileEntry]()
     }
   }
 
+  private def getUserFileRecord(user: User): util.List[DashboardFileEntry] = {
+    val accesses = userFileAccessDao.fetchByUid(user.getUid)
+    val fileEntries: mutable.ArrayBuffer[DashboardFileEntry] = mutable.ArrayBuffer()
+    accesses.asScala.toList.map(access => {
+      val fid = access.getFid
+      val file = fileDao.fetchOneByFid(fid)
+      var accessLevel = "None"
+      if (access.getWriteAccess) {
+        accessLevel = "Write"
+      } else if (access.getReadAccess) {
+        accessLevel = "Read"
+      } else {
+        accessLevel = "None"
+      }
+      val ownerName = userDao.fetchOneByUid(file.getUid).getName
+      fileEntries += DashboardFileEntry(
+        ownerName,
+        accessLevel,
+        ownerName == user.getName,
+        file
+      )
+    })
+    fileEntries.toList.asJava
+  }
+
+  /**
+    * This method deletes a file from a user's repository
+    * @param fileName the name of file being deleted
+    * @param ownerName the name of the file's owner
+    * @param session the session indicating the current user
+    * @return
+    */
   @DELETE
-  @Path("/delete/{fileID}")
+  @Path("/delete/{fileName}/{ownerName}")
   def deleteUserFile(
-      @PathParam("fileID") fileID: UInteger,
+      @PathParam("fileName") fileName: String,
+      @PathParam("ownerName") ownerName: String,
       @Session session: HttpSession
   ): Response = {
 
     UserResource.getUser(session) match {
       case Some(user) =>
+        val fileID = UserFileAccessResource.getFileId(ownerName, fileName)
         val userID = user.getUid
-        // TODO: add user check
-        val filePath = fileDao.fetchOneByFid(fileID).getPath
-        UserFileUtils.deleteFile(Paths.get(filePath))
-        fileDao.deleteById(fileID)
-        Response.ok().build()
+        val hasWriteAccess = context
+          .select(USER_FILE_ACCESS.WRITE_ACCESS)
+          .from(USER_FILE_ACCESS)
+          .where(USER_FILE_ACCESS.UID.eq(userID).and(USER_FILE_ACCESS.FID.eq(fileID)))
+          .fetch()
+          .getValue(0, 0)
+        if (hasWriteAccess == false) {
+          Response
+            .status(Response.Status.UNAUTHORIZED)
+            .entity("You do not have the access to deleting the file")
+            .build()
+        } else {
+          val filePath = fileDao.fetchOneByFid(fileID).getPath
+          UserFileUtils.deleteFile(Paths.get(filePath))
+          fileDao.deleteById(fileID)
+          Response.ok().build()
+        }
+
       case None =>
         Response.status(Response.Status.UNAUTHORIZED).build()
     }
@@ -121,12 +200,6 @@ class UserFileResource {
 
   }
 
-  private def getUserFileRecord(userID: UInteger): util.List[File] = {
-
-    // TODO: verify user in session?
-    fileDao.fetchByUid(userID)
-  }
-
   private def validateFileName(fileName: String, userID: UInteger): Pair[Boolean, String] = {
     if (fileName == null) Pair.of(false, "file name cannot be null")
     else if (fileName.trim.isEmpty) Pair.of(false, "file name cannot be empty")
@@ -135,8 +208,8 @@ class UserFileResource {
   }
 
   private def isFileNameExisted(fileName: String, userID: UInteger): Boolean =
-    SqlServer.createDSLContext.fetchExists(
-      SqlServer.createDSLContext
+    context.fetchExists(
+      context
         .selectFrom(FILE)
         .where(FILE.UID.equal(userID).and(FILE.NAME.equal(fileName)))
     )
